Simplify server tab bar by throwing out parallel routing
