+++ app/stats/MetricBox.tsx
