+++ app/stats/foo/page.tsx
